+++ packages/cli/tsconfig.base.json
@@ -29,5 +29,5 @@
     "rootDir": "./src",
     "outDir": "./dist"
   },
-  "include": ["./src/**/*.ts", "./test/**/*.ts"]
+  "include": ["./src/**/*.ts"]
 }
